PlatformServiceを作成し、GameServiceに記述したPlatformの内容をPlatformServiceに移した
